feat: create PaymentMethodResources
